Change unique_fields check against db to ensure schema matches LicenceRecord
Sqlite’s insert_or_update method is problematic when the unique indexes in the schema change unexpectedly. If a requirement for a unique index is added at the point of database creation but then later removed by the developer, this can lead to a situation where records are duplicated instead of updated in place. Combined with the Enumerator aspects of the bot, this can lead to dangerous loops when exporting the data.

This patch ensures that the unique_fields method matches the unique indexes in the db schema on every run.