Finish consolidating database queries into one place for Simple Bot
The current thinking behind this is to offer potential for a “default_scope” over a bot’s database. In some cases we might be collecting a mix of record types but only want to test and export the Finance related records for example. This commit simplifies the number of ways the database is queried to run via all_stored_records for now