Improve spec
In fact the expectation is that `puts` is never called.

When `puts` *is* called it is with this data but JSONified in a different key order so the strings are not in fact the same.

In other words this spec was testing for absence of a particular string that never would have arrived anyway when it should have been testing for the absence of the method call. It passed by luck rather than because it tested what needed to be tested.