Add and configure rubocop gem
`bundle install && bundle binstubs rubocop`

For the configuration we included the settings used habitually by the OpenCorporates tech team in the belief that they are rational.

The `Style/EmptyMethod` and `Style/TrailingCommaIn*` cops are set so that diffs are easy to understand when a developer adds a new line of code.

The `Style/StringLiterals` setting is to ensure a consistent style for something that Ruby developers have never achieved consensus on[citation needed].