Tighten up version declarations
We don't want to allow the next major versions of these dependencies in case the gem is not compatible with them. The versions mentioned are the latest current version of each gem.

If an app that includes an older version of this gem later upgrades gems mentioned here to new major versions, there is the possibility that they will have to downgrade them again when they update this gem. BUT, we need to ensure that this gem is only used in an environment where the code it depends on it is known to be compatible.