Silence stdout logging during testing
When in test mode StatsD wants to log to stdout by default. Silence that to keep rspec output clean.

TODO: We might want to think about logging to a file, but maybe not for just this statsd output. If we introduced a bot-wide logger object, then we would include this output in there.