Provide subject_details option on send_error_report
Provide a new `:subject_details` option on the `send_error_report` method, allowing a bot developer to override the default (simple exception message) output.

This will allow us (as and when we are ready to do so), to blow away of a lot of messy code in our bots, where we override (and largely duplicate) the send_error_report method.