Introduce @processed_count instance var
Introduce the `increment_progress_counters` method which sets and increments a @processed_count instance variable. This works as a way of allowing a gem_only change to do progress reporting with state of the counters managed in this gem module.

For now we will change the reporting method to reflect that we only know a "companies_processed" stat.