Add host to run report
Add a new field host to the final run report. Analysis app will add this to the "output" display. It may be useful to distinguish manual ad-hoc or development runs from those which run on backgrounder2.
Only proper runs should be reported, since only those *should* be run with `FETCHER_BOT_ENV=production`, but this is a precaution to help understand what is happening in case this is not the case.