Warn when missing FETCHER_BOT_ENV on other tasks
Add a warning to the top of the output for other tasks.
In future we want to distinguish whether or not we are running any of these rake tasks "in production". While it currently makes no difference, we should figure out when and where these rake tasks are invoked, and try to always set FETCHER_BOT_ENV accordingly.