Remove global methods from misleading namespace
Remove (and unindent) various methods out of the `namespace` block where one might be mislead into believing that they are namespaced. They are not. Only the task definitions are namespaced. (If we wanted these methods to not pollute the global namespace we would need to define a module)

This also reduces the length of our max block size rubocop metric.