Call run from the bot:run task
Now that we are disabling reporting if FETCHER_BOT_ENV=development, we can call the `run` method from both `bot:run2` and `bot:run` tasks. It will skip reporting as required on a call to `bot:run`.

TODO: This is a rationalisation which paves the way for simply doing away with `bot:run2` in future, but that would assume we are in the habit of always setting FETCHER_BOT_ENV correctly. For now we retain the two tasks.