Fix statsd jurisdiction codes to include underscores
Call the `inferred_jurisdiction_code` method, which is defined in `CompanyFetcherBot`, to produce valid statsd_namespace and jurisdiction_code in snake_case format.

In practice that method is always available (since all bots are company fetcher bots) but in theory OpencBot::CompanyFetcherBot is optional, and in tests e.g. of the reporting helper where we test out statsd, it is not implemented, hence a separate test within company_fetcher_bot_spec.