OpencBot configuration from Analysis App
This commit allows a fetcher bot to retrieve information about the bot from Analysis App's bots.json output. We will need the jurisdiction_code to be able to pull the data from the JSON. I considered using a Proc select method of getting this data, as it reads nicer, bu this is slightly more performative.