Deprecate unused definition of update_data
Begin the process of removing `RegisterMethods.update_data` which is presumed to be dead code. `update_data` is a very important top level method, the active definition of which can be found in company_fetcher_bot.rb. It is not clear why it is also defined in the register_methods.rb helper where it will (presumably) always get overwritten by company_fetcher_bot.rb.

It can therefore be deleted. We will do so at some point in the future when we are sure nothing hits this error message. This will remove a tiny bit of confusing dead code, but it will also mean can delete save_run_report.