Assess the stale count in the sqlite
This will append the stale count to be sent in the end of run report.
Update stale is called as part of update_data in company_fetcher_bot and the data that is returned here is included in the final run report, which is why we are returning the stale count here.
There was a question around the performance of this query and how it would effect bot runs, but the finalized reporting is done when the bot has finished its processing, so this will cause a slight delay in report time, but not effect freshness metrics.
The update_stale spec test requires the before_do save_data to create the stub sqlite to query for the stale metric.
The days till stale method is used as the defining factor for the metric as we wish to mirror the stale_entry_uids flow to analyse how many records the fetcher will be rotating through on its next run, with some fall over allowed for the next day.