Add resque gem and set-up
Add the resque gem as a dependency, and initialise resque redis connection within the rake task config.

It will look for a yaml config file at the path specified by EXTERNAL_BOTS_RESQUE_CONFIG, or in the external_bots/config/resque.yml . Of course it needs a connection to the openc redis instance for message passing to work, so the contents of this file, including prod sentinel settings, will be identical, or we'll just link to the file in the openc config.