Add resque worker classes
Add resque worker class SruRequestJob with `perform` method where the actual invoking of a single record update happens (call to `update_datum`). The response data is returned as JSON, including logic to catch any excpetion and return it as JSON. That is the same as the previous rake mechanism, apart from we no longer puts it stdout for ingestion in a docker ssh session. We now just put the json into the 'output' field of our reply resque message.

For the resque reply message, we add a SingleRecordUpdateJob class and `enqueue` method. Instances of this class are passed back on the queue as a response back to openc, where the _real_ SingleRecordUpdateJob class (with `perform` method) is defined.