Enable NewCops and regenerate todo file
With rubocop upgrades we get lots of warnings about unconfigured cops. We just can just activate them with `NewCops: enable` and then simply regenerate the todo file to list any failures of the new cops (on this occasion and any time we upgrade rubocop in future)

Note: With the previous `bundle upgrade` we also brought new cops of the rubocop-rspec gem.