Prioritise active when updating stale companies
When retrieving stale companies to update, split this into two different queries: active stale and inactive stale. Process the two resultsets in an interleaved fashion, to acheive the desired active/inactive ratio.

The desired ratio is configured in the bot code as ACTIVE_RATIO.

The total number of updates will still be STALE_COUNT (assuming there are more than this number remaining to update).